requirements.txt
opencv-python
numpy
pyzbar
pyserial
sounddevice
fpdf2
requests
pycryptodomex